+++ src/Plivo/Resource/Profile/ProfileInterface.cs
@@ -124,7 +124,7 @@ namespace Plivo.Resource.Profile
 
         public ProfileResponse Create(string profile_alias, string plivo_subaccount,  string customer_type,
             string entity_type, string company_name,  string ein, string vertical,
-            string ein_issuing_country, string stock_symbol, string stock_exchange, string website,string alt_business_id, string alt_business_id_type, AuthorizedContact authorized_contact, Address address, string business_contact_email = null, string doing_business_as = null)
+            string ein_issuing_country, string stock_symbol, string stock_exchange, string website,string alt_business_id, string alt_business_id_type, AuthorizedContact authorized_contact, Address address, string business_contact_email = null, string doing_business_as = null, string number_of_employees = null)
         {
         var mandatoryParams = new List<string>{"profile_alias"};
         Console.WriteLine(entity_type);
@@ -148,7 +148,8 @@ namespace Plivo.Resource.Profile
                 address,
                 website,
                 business_contact_email,
-                doing_business_as
+                doing_business_as,
+                number_of_employees
             });
 		return ExecuteWithExceptionUnwrap(() =>
 		{
@@ -179,7 +180,7 @@ namespace Plivo.Resource.Profile
         /// <param name="business_contact_email">business_contact_email</param>
 		public async Task<ProfileResponse> CreateAsync(string profile_alias, string plivo_subaccount,  string customer_type,
             string entity_type, string company_name,  string ein, string vertical, string website,
-            string ein_issuing_country, string stock_symbol, string stock_exchange, string alt_business_id, string alt_business_id_type, AuthorizedContact authorized_contact, Address address, string business_contact_email = null, string doing_business_as = null)
+            string ein_issuing_country, string stock_symbol, string stock_exchange, string alt_business_id, string alt_business_id_type, AuthorizedContact authorized_contact, Address address, string business_contact_email = null, string doing_business_as = null, string number_of_employees = null)
 		{
          var mandatoryParams = new List<string>{"profile_alias"};
           var data = CreateData(
@@ -202,7 +203,8 @@ namespace Plivo.Resource.Profile
                 address,
                 website,
                 business_contact_email,
-                doing_business_as
+                doing_business_as,
+                number_of_employees
             });
 			var result = await Client.Update<ProfileResponse>(Uri + "Profile/", data);
             result.Object.StatusCode = result.StatusCode;
@@ -227,7 +229,7 @@ namespace Plivo.Resource.Profile
 
         public GetProfile Update(string profile_uuid, string company_name =null,  string website = null,
             string entity_type=null, string vertical= null,  AuthorizedContact authorized_contact=null, Address address=null, string business_contact_email = null,
-            string ein = null, string ein_issuing_country = null, string alt_business_id = null, string alt_business_id_type = null, string doing_business_as = null)
+            string ein = null, string ein_issuing_country = null, string alt_business_id = null, string alt_business_id_type = null, string doing_business_as = null, string number_of_employees = null)
         {
              var mandatoryParams = new List<string>{"profile_uuid"};
         var data = CreateData(
@@ -245,7 +247,8 @@ namespace Plivo.Resource.Profile
                 ein_issuing_country,
                 alt_business_id,
                 alt_business_id_type,
-                doing_business_as
+                doing_business_as,
+                number_of_employees
             });
 		return ExecuteWithExceptionUnwrap(() =>
 		{
@@ -270,7 +273,7 @@ namespace Plivo.Resource.Profile
         /// <param name="business_contact_email">business_contact_email</param>
 		public async Task<ProfileResponse> UpdateAsync(string profile_uuid, string company_name =null,  string website = null,
             string entity_type=null, string vertical= null,  AuthorizedContact authorized_contact=null, Address address=null, string business_contact_email = null,
-            string ein = null, string ein_issuing_country = null, string alt_business_id = null, string alt_business_id_type = null, string doing_business_as = null)
+            string ein = null, string ein_issuing_country = null, string alt_business_id = null, string alt_business_id_type = null, string doing_business_as = null, string number_of_employees = null)
         {
              var mandatoryParams = new List<string>{"profile_uuid"};
         var data = CreateData(
@@ -288,7 +291,8 @@ namespace Plivo.Resource.Profile
                 ein_issuing_country,
                 alt_business_id,
                 alt_business_id_type,
-                doing_business_as
+                doing_business_as,
+                number_of_employees
             });
 			var result = await Client.Update<ProfileResponse>(Uri + "Profile/"+profile_uuid+"/", data);
             result.Object.StatusCode = result.StatusCode;